Integrate Razorpay Checkout Form
This commit enables the razorpay checkout form.
The razorpay checkout form is initialized from the payment page and allows an user to make a payment by using Razorpay.
The order and payment on razorpay's end are created.
This commit allows the user to make a successful checkout by using the Razorpay Payment Method.