[FIX] Incorrect Status implementation on payment_source
This commit fixes an issue with the payment source where status was not being recorded.
The issue was fixed by removing enums defined on the model and changing the type of the column in the migration from integer to string.
Columns affected in the table solidus_razorpay_payment_sources are:
- status
- refund_status