Refund API Implementation
This commit implements 2 methods on the gateway.rb file to enable refunds.
The methods implemented are:
- void
- credit
The void method cancels a payment and refunds the amouny. It is called when the admin uses the void option in order payments.
The credit method creates a refund for the order. It is called when admin uses the refund option in order payments.