Modify checkout helper to use exisitng razorpay orders
This commit modifies the checkout_helper to use an existing razorpay_order

Earlier behaviour created a new razorpay_order for every payment request which was wrong.
Now whenever a payment request is made, the spree_order is checked if it has an existing razorpay_order.
If a razorpay_order exists, then payments are made against the existing order, else a new razorpay_order is created.